+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
+- Fixed a bug in generating the migration script when data migration with deleting the identity column of the table.
 - Fixed display of object references in Outline for MSSQL.
 
 ## [7.4.0] - 2023-03-28